Split direct build validation fixtures
