+++ tests/integration/cli_build/direct_build_graph_validation.rs
@@ -1,9 +1,13 @@
 use std::process::Command;
 
+#[path = "direct_build_graph_validation/dependency_shapes.rs"]
+mod dependency_shapes;
 #[path = "direct_build_graph_validation/gated_dependencies.rs"]
 mod gated_dependencies;
 #[path = "direct_build_graph_validation/graph_only_libraries.rs"]
 mod graph_only_libraries;
+#[path = "direct_build_graph_validation/target_metadata.rs"]
+mod target_metadata;
 
 #[test]
 fn direct_file_command_build_zen_rejects_graph_without_executable_targets() {
@@ -105,191 +109,6 @@ main = () i32 {
     assert!(
         tmp.path().join("build").join("app").join("app").exists(),
         "expected executable output to exist"
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_unknown_target_dependencies() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["core"],
-    })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    assert_direct_file_command_rejects_dependency_shape(
-        tmp.path(),
-        "build target `app` depends on unknown target `core`",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_self_target_dependencies() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["app"],
-    })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    assert_direct_file_command_rejects_dependency_shape(
-        tmp.path(),
-        "build target `app` cannot depend on itself",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_cyclic_target_dependencies() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["tool"],
-    })
-    b.add(Executable {
-        name: "tool",
-        main: "tool.zen",
-        out_dir: "build/tool/",
-        dependencies: ["app"],
-    })
-    .Ok(b.config())
-}
-"#,
-    )
-    .expect("write build.zen");
-    assert_direct_file_command_rejects_dependency_shape(
-        tmp.path(),
-        "build target dependency cycle includes `app`",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_duplicate_target_fields() {
-    assert_direct_file_command_rejects_target_metadata(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        name: "tool",
-        main: "app.zen",
-        out_dir: "build/app/",
-    })
-    .Ok(b.config())
-}
-"#,
-        "duplicate field `name` in `Executable` build target",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_missing_required_target_fields() {
-    assert_direct_file_command_rejects_target_metadata(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-    })
-    .Ok(b.config())
-}
-"#,
-        "missing required field `out_dir` in `Executable` build target",
-    );
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_invalid_target_field_types() {
-    assert_direct_file_command_rejects_target_metadata(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: 42,
-    })
-    .Ok(b.config())
-}
-"#,
-        "field `out_dir` in `Executable` build target must be a string",
-    );
-}
-
-fn assert_direct_file_command_rejects_target_metadata(
-    build_source: &str,
-    expected_diagnostic: &str,
-) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(tmp.path().join("build.zen"), build_source).expect("write build.zen");
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .arg("build.zen")
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build.zen");
-
-    assert!(
-        !output.status.success(),
-        "zen build.zen unexpectedly succeeded: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains(expected_diagnostic),
-        "expected target metadata diagnostic `{expected_diagnostic}`, stderr={stderr}"
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "direct build.zen command should not create outputs after target metadata validation fails"
-    );
-}
-
-fn assert_direct_file_command_rejects_dependency_shape(
-    project_dir: &std::path::Path,
-    expected_diagnostic: &str,
-) {
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .arg("build.zen")
-        .current_dir(project_dir)
-        .output()
-        .expect("run zen build.zen");
-
-    assert!(
-        !output.status.success(),
-        "zen build.zen unexpectedly succeeded: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stderr).contains(expected_diagnostic),
-        "expected dependency diagnostic `{expected_diagnostic}`, stderr={}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        !project_dir.join("build").exists(),
-        "direct build.zen command should not create outputs after dependency validation fails"
     );
 }
